Issue 3012: playing with AUGraphicEQ didn't apply initial sliders...

... this surprisingly complicated fix had many parts involving

- guaranteeing the initial population of EffectSettings
- fixing the omission of re-fetch of settings in the event handler, storing of parameter
- names in EffectSettings (because they aren't always just metadata
but can depend on other settings; slider names depend on the 10/31
choice)
- two passes when storing settings (because whether one control is
assignable may depend on the state of another control)

And there are also some improvements of error handling in the
general framework of realtime effects.

+++ src/effects/StatefulEffectBase.cpp
@@ -86,10 +86,12 @@ bool StatefulEffectBase::Instance::RealtimeProcessStart(EffectSettings &settings
    return GetEffect().RealtimeProcessStart(settings);
 }
 
-size_t StatefulEffectBase::Instance::RealtimeProcess(int group, EffectSettings &settings,
+size_t StatefulEffectBase::Instance::RealtimeProcess(size_t group,
+   EffectSettings &settings,
    const float *const *inBuf, float *const *outBuf, size_t numSamples)
 {
-   return GetEffect().RealtimeProcess(group, settings, inBuf, outBuf, numSamples);
+   return GetEffect()
+      .RealtimeProcess(group, settings, inBuf, outBuf, numSamples);
 }
 
 bool StatefulEffectBase::Instance::RealtimeProcessEnd(EffectSettings &settings) noexcept
@@ -154,7 +156,8 @@ bool StatefulEffectBase::RealtimeProcessStart(EffectSettings &settings)
    return true;
 }
 
-size_t StatefulEffectBase::RealtimeProcess(int group, EffectSettings &settings,
+size_t StatefulEffectBase::RealtimeProcess(size_t group,
+   EffectSettings &settings,
    const float *const *inbuf, float *const *outbuf, size_t numSamples)
 {
    return 0;
