+++ src/effects/StatefulEffectBase.h
@@ -39,7 +39,7 @@ public:
       bool RealtimeSuspend() override;
       bool RealtimeResume() noexcept override;
       bool RealtimeProcessStart(EffectSettings &settings) override;
-      size_t RealtimeProcess(int group, EffectSettings &settings,
+      size_t RealtimeProcess(size_t group, EffectSettings &settings,
          const float *const *inBuf, float *const *outBuf, size_t numSamples)
       override;
       bool RealtimeProcessEnd(EffectSettings &settings) noexcept override;
@@ -101,7 +101,7 @@ public:
      @copydoc RealtimeInitialize::RealtimeProcess()
      Default implementation does nothing, returns 0
    */
-   virtual size_t RealtimeProcess(int group, EffectSettings &settings,
+   virtual size_t RealtimeProcess(size_t group, EffectSettings &settings,
       const float *const *inBuf, float *const *outBuf, size_t numSamples);
 
    /*!
